# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: oneSCM Handbook Studio API
  version: 0.12.0
  description: |
    Revisionssichere Erstellung rollen- und spartenspezifischer Benutzerhandbücher aus Markdown-Quellen.
    Jede Operation referenziert über `x-requirements` die zugehörigen User Stories (US-020).
    Operationen mit `x-extension: true` ergänzen die verbindliche Operationsliste aus Masterprompt §10 (siehe docs/01-gap-analyse.md, Lücke 10).
    Fehler werden als `application/problem+json` (RFC 9457) geliefert.
    Authentifizierung (ENTSCHEIDUNG E-15), je nach Betriebsart (`GET /auth/config`):
    - `oidc`: `Authorization: Bearer <JWT>` eines OpenID-Connect-Providers; technische Berechtigungen aus einem Gruppen-/Rollen-Claim.
    - `demo`: Header `X-User-Id` mit einem Demo-Benutzer; ohne Header nur Lesezugriff. Nur für Entwicklung und Demo.
    Projekt (ADR-014): Header `X-Project-Id` wählt das Projekt aller projektbezogenen Endpunkte (Standard `p_default`);
    IDs anderer Projekte werden wie nicht vorhandene behandelt (404). Ausnahmen: `/projects…` und die Health-Checks.
    Betrieb (ADR-015, ADR-027): jede Antwort trägt `X-Request-Id` (bei aktivem Tracing zusätzlich `X-Trace-Id`, W3C `traceparent` wird übernommen);
    Rate-Limits je Benutzer und Minute (429 mit `Retry-After`), mit `RATE_LIMIT_STORE=db` gemeinsam über alle Instanzen;
    Prometheus-Metriken unter `/metrics` (außerhalb von `/api/v1`, Bearer `METRICS_TOKEN` oder Berechtigung admin).
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-Token des konfigurierten OIDC-Providers (AUTH_MODE=oidc)
    demoUser:
      type: apiKey
      in: header
      name: X-User-Id
  parameters:
    ImportId: { name: importId, in: path, required: true, schema: { type: string } }
    SnippetId: { name: snippetId, in: path, required: true, schema: { type: string } }
...
  responses:
...
    Unauthorized:
      description: Nicht angemeldet oder Token ungültig
      headers:
        WWW-Authenticate: { schema: { type: string, example: Bearer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem:
      description: Fehler (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        gate: { $ref: '#/components/schemas/GateResult' }
        blockers: { type: array, items: { type: object } }
    RoleCode: { type: string, enum: [dealer, market, mo, hq, all] }
    DivisionCode: { type: string, enum: [car, van, truck, bus, all, unconfirmed] }
    EvidenceStatus: { type: string, enum: [source_confirmed, manually_confirmed, unconfirmed, open_question] }
    Assignment:
      type: object
      properties:
        code: { type: string }
        score: { type: number }
        method: { type: string }
        modelVersion: { type: string }
        evidenceStatus: { $ref: '#/components/schemas/EvidenceStatus' }
        evidence: { type: [string, 'null'] }
    Import:
      type: object
      properties:
        id: { type: string }
        fileName: { type: string }
        kind: { type: string, enum: [zip, md] }
        sha256: { type: string }
        byteSize: { type: integer }
        status: { type: string, enum: [queued, processing, completed, completed_with_errors, failed] }
        createdAt: { type: string, format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        stats: { type: object }
        items:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              sha256: { type: [string, 'null'] }
              status: { type: string, enum: [imported, identical, failed, skipped] }
              message: { type: [string, 'null'] }
              revisionId: { type: [string, 'null'] }
    Snippet:
      type: object
      properties:
        id: { type: string }
        seq: { type: integer }
        text: { type: string, description: Unveränderlicher Ursprungstext }
        kind: { type: string, enum: [paragraph, list, ordered_list, table, code, quote] }
        chapter: { type: object, properties: { id: { type: string }, title: { type: string } } }
        subchapter: { type: [object, 'null'] }
        headingPath: { type: array, items: { type: string } }
        lineStart: { type: integer }
        lineEnd: { type: integer }
        evidenceStatus: { $ref: '#/components/schemas/EvidenceStatus' }
        market: { type: [string, 'null'] }
        release: { type: [string, 'null'] }
        scopeStatus: { type: string }
        excludedReason: { type: [string, 'null'] }
        source: { type: object }
        roles: { type: array, items: { $ref: '#/components/schemas/Assignment' } }
        divisions: { type: array, items: { $ref: '#/components/schemas/Assignment' } }
    SnippetPatch:
      type: object
      description: Ändert ausschließlich Klassifikation und Evidenzstatus, nie den Ursprungstext.
      properties:
        roles: { type: array, items: { $ref: '#/components/schemas/RoleCode' } }
        divisions: { type: array, items: { $ref: '#/components/schemas/DivisionCode' } }
        confirmRoles: { type: boolean }
        confirmDivisions: { type: boolean }
        evidenceStatus: { $ref: '#/components/schemas/EvidenceStatus' }
        market: { type: [string, 'null'] }
        release: { type: [string, 'null'] }
        confirmScope: { type: boolean }
        note: { type: [string, 'null'] }
...
    GateResult:
      type: object
      properties:
        passed: { type: boolean }
        checks:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              label: { type: string }
              passed: { type: boolean }
              details: { type: array, items: { type: string } }
...
security:
  - oidcBearer: []
  - demoUser: []
paths:
  /imports:
    post:
      operationId: createImport
      summary: ZIP-, Markdown-, HTML- (Confluence-Export) oder Word-Datei importieren (HTML/Word werden in Markdown umgewandelt, ADR-022)
      tags: [Import]
      x-requirements: [US-001, US-002, US-003, US-004, NFR-09, NFR-14, NFR-23]
      parameters:
        - { name: snapshot, in: query, schema: { type: boolean }, description: 'ZIP ist der vollständige Stand – fehlende Dateien derselben Herkunft werden als entfernt markiert (ADR-036)' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '202':
          description: Import angenommen, Verarbeitung im Hintergrund
          content: { application/json: { schema: { $ref: '#/components/schemas/Import' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '415': { $ref: '#/components/responses/Problem' }
    get:
      operationId: listImports
      summary: Importe auflisten
      tags: [Import]
      x-requirements: [US-001]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Liste, content: { application/json: { schema: { type: array, items: { $ref: ''#/components/schemas/Import'' } } } }' }
  /imports/{importId}:
    get:
      operationId: getImport
      summary: Importstatus und Protokoll je Datei
      tags: [Import]
      x-requirements: [US-001]
      parameters: [{ $ref: '#/components/parameters/ImportId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Import, content: { application/json: { schema: { $ref: ''#/components/schemas/Import'' } } }' }
        '404': { $ref: '#/components/responses/Problem' }
  /sources:
    get:
      operationId: listSources
      summary: Quelldokumente mit Revisionen (removedAt – als entfernt markiert, ADR-036)
      tags: [Import]
      x-requirements: [US-001, US-011]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Quellen }
  /source-documents/{documentId}/restore:
    post:
      operationId: restoreSourceDocument
      summary: 'Als entfernt markiertes Quelldokument wiederherstellen (letzte Revision wird wieder aktuell)'
      tags: [Import]
      x-requirements: [NFR-23]
      x-extension: true
      parameters: [{ name: documentId, in: path, required: true, schema: { type: string } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Quelldokument' }
        '409': { description: 'nicht als entfernt markiert' }
  /source-revisions/{revisionId}/raw:
    get:
      operationId: getRevisionRaw
      summary: Unveränderten Ursprungstext einer Revision als text/plain abrufen
      tags: [Import]
      x-requirements: [US-001, US-011]
      x-extension: true
      parameters: [{ name: revisionId, in: path, required: true, schema: { type: string } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Rohtext, content: { text/plain: { schema: { type: string } } }' }
        '404': { $ref: '#/components/responses/Problem' }
  /source-revisions/{revisionId}/original:
    get:
      operationId: getRevisionOriginal
      summary: Originaldatei einer umgewandelten Revision (HTML als text/plain, Word als .docx; ADR-022)
      tags: [Import]
      x-requirements: [NFR-09, US-001]
      x-extension: true
      parameters: [{ name: revisionId, in: path, required: true, schema: { type: string } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Originaldatei }
        '404': { description: Revision unbekannt oder Markdown-Quelle ohne Original }
...
  /snippets:
    get:
      operationId: listSnippets
      summary: Textabschnitte suchen und kombiniert filtern
      tags: [Snippets]
      x-requirements: [US-002, US-003, US-004, US-010, US-017]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: chapterId, in: query, schema: { type: string } }
        - { name: subchapterId, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { $ref: '#/components/schemas/RoleCode' } }
        - { name: division, in: query, schema: { $ref: '#/components/schemas/DivisionCode' } }
        - { name: market, in: query, schema: { type: string } }
        - { name: release, in: query, schema: { type: string } }
        - { name: evidenceStatus, in: query, schema: { $ref: '#/components/schemas/EvidenceStatus' } }
        - { name: findingType, in: query, schema: { type: string } }
        - { name: includeHistoric, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Seite
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
                  items: { type: array, items: { $ref: '#/components/schemas/Snippet' } }
  /snippets/{snippetId}:
    get:
      operationId: getSnippet
      summary: Einzelnen Textabschnitt mit Quelle und Klassifikation abrufen
      tags: [Snippets]
      x-requirements: [US-002, US-011]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/SnippetId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Textabschnitt, content: { application/json: { schema: { $ref: ''#/components/schemas/Snippet'' } } }' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: patchSnippet
      summary: Rollen, Sparten, Markt, Release und Evidenzstatus bestätigen oder ändern
      tags: [Snippets]
      x-requirements: [US-003, US-004]
      parameters: [{ $ref: '#/components/parameters/SnippetId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SnippetPatch' } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Aktualisiert, content: { application/json: { schema: { $ref: ''#/components/schemas/Snippet'' } } }' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /settings:
    get:
      operationId: getSettings
      summary: Einstellungen (Schwellenwerte, Uploadgrenzen, Terminologie)
      tags: [System]
      x-requirements: [US-005, US-001, US-015]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Einstellungen }
    put:
      operationId: putSettings
      summary: Einstellungen ändern (Berechtigung admin)
      tags: [System]
      x-requirements: [US-005, US-001, US-015]
      x-extension: true
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Einstellungen }
        '403': { $ref: '#/components/responses/Problem' }
...
  /media:
    get:
      operationId: listMedia
      summary: Bilder des Projekts (inhaltsadressiert, ADR-029)
      tags: [Medien]
      x-requirements: [NFR-16]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'SHA-256, Format, Größe, Abmessungen, Herkunft' }
    post:
      operationId: uploadMedia
      summary: 'Bild hochladen (PNG, JPEG, GIF, WebP nach Dateisignatur; SVG wird aus einer Positivliste neu geschrieben, ADR-036; Berechtigung edit); Antwort enthält den Markdown-Verweis ![…](media:<sha256>)'
      tags: [Medien]
      x-requirements: [NFR-16, NFR-23]
      x-extension: true
      requestBody:
        content: { multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: Abgelegt }
        '400': { description: Kein unterstütztes Bildformat }
        '413': { description: Zu groß }
  /media/{mediaSha}:
    get:
      operationId: getMedia
      summary: Bild ausliefern (nur eigenes Projekt, nosniff, unveränderlich zwischenspeicherbar)
      tags: [Medien]
      x-requirements: [NFR-16]
      x-extension: true
      parameters: [{ name: mediaSha, in: path, required: true, schema: { type: string, pattern: '^[a-f0-9]{64}$' } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Bilddaten, content: { image/png: {}, image/jpeg: {}, image/gif: {}, image/webp: {} }' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: setMediaTitle
      summary: Bildtitel für das Bildverzeichnis setzen (Berechtigung edit)
      tags: [Stammdaten]
      x-requirements: [NFR-19, NFR-16]
      x-extension: true
      parameters: [{ name: mediaSha, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { title: { type: [string, 'null'] } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: OK }
...